requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
lxml==4.9.2
pytz==2023.2
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
